requirements.txt
numpy
sympy
roboticstoolbox-python
spatialmath-python